Fix silent errors
* report error from within threads
* fix issue when object size in header is larger than actual object read

+++ src/Vector/BLF/UncompressedFile.cpp
@@ -201,7 +201,8 @@ void UncompressedFile::write(const std::shared_ptr<LogContainer> & logContainer)
     tellgChanged.wait(lock, [&] {
         return
         m_abort ||
-        static_cast<uint32_t>(m_tellp - m_tellg) < m_bufferSize;
+        /* m_tellg can get larger if using seekg */
+        ((m_tellp - m_tellg) < m_bufferSize);
     });
 
     /* append logContainer */
